+++ doc/Changelog.md
@@ -1,8 +1,11 @@
 Changelog {#Changelog}
 =========
 
-# Release 2.0.0 (24-May-2017)
+# Release 2.0.0 (30-May-2017)
 
+* [157](https://github.com/BlueBrain/Brion/pull/157):
+  - Add Synapse.gid() python wrapping
+  - Improve synapse attribute reading performance
 * [154](https://github.com/BlueBrain/Brion/pull/154):
   - Improved help and command line inverface for spikeConverter and
     compartmentConverter.